Issue number: PCN20707 - Allow specification of an artifact type as for use by the application, and therefore immutable from the admin ui.
#getAssociatedIssueTypes()-New method added to get the list of issue types an
attribute is associated with.
#isSystemDefined()-New method added to check if the attribute is associated with
a system specified issue type.

src/java/org/tigris/scarab/om/IssueType.java
#isSystemDefined()-New method added to check if it is a system specified issue type.

src/java/org/tigris/scarab/actions/admin/GlobalArtifactTypeCreate.java
src/java/org/tigris/scarab/actions/admin/GlobalAttributeEdit.java
src/java/org/tigris/scarab/actions/admin/IssueTypeAttributeEdit.java
src/java/org/tigris/scarab/actions/admin/ArtifactTypeEdit.java
src/java/org/tigris/scarab/actions/admin/AttributeGroupEdit.java
src/java/org/tigris/scarab/actions/admin/GlobalArtifactTypes.java
src/templates/screens/admin/GlobalArtifactTypeEdit.vm
src/templates/screens/admin/GlobalAttributeGroupEdit.vm
src/templates/screens/admin/GlobalAttributeEdit.vm
src/templates/screens/admin/UserAttributeEdit.vm
src/templates/screens/admin/ArtifactTypeEdit.vm
src/templates/screens/admin/AttributeGroupEdit.vm
src/templates/screens/admin/GlobalArtifactTypeShow.vm
#In all the above files, it is checked if the issue type/attribute is system
defined and if so, actions related to deleting/editing are prevented.

src/i18n/ScarabBundle_en_US.properties
#New resources added to alert when the user attempts to edit system defined
artifact types.

Submitted by: venkat@collab.net
Reviewed by: jmcnally@collab.net

+++ src/java/org/tigris/scarab/actions/admin/GlobalAttributeEdit.java
@@ -77,7 +77,7 @@ import org.tigris.scarab.services.cache.ScarabCache;
  * This class deals with modifying Global Attributes.
  *
  * @author <a href="mailto:jon@collab.net">Jon S. Stevens</a>
- * @version $Id: GlobalAttributeEdit.java,v 1.61 2003/07/25 17:18:00 thierrylach Exp $
+ * @version $Id: GlobalAttributeEdit.java,v 1.62 2003/07/26 18:26:57 jmcnally Exp $
  */
 public class GlobalAttributeEdit extends RequireLoginFirstAction
 {
@@ -182,10 +182,16 @@ public class GlobalAttributeEdit extends RequireLoginFirstAction
     {
         ScarabRequestTool scarabR = getScarabRequestTool(context);
         Attribute attr = scarabR.getAttribute();
+        if (attr.isSystemDefined())
+        {
+            scarabR.setAlertMessage(getLocalizationTool(context).get
+	                              ("CannotDeleteSystemSpecifiedAttribute"));
+            return;
+        }
         if (attr.getAttributeId() != null)
         {
             attr.deleteModuleMappings(); 
-            attr.deleteIssueTypeMappings(); 
+            attr.deleteIssueTypeMappings();
             attr.setDeleted(true);
             attr.save();
             scarabR.setConfirmMessage(getLocalizationTool(context).get(DEFAULT_MSG));  
@@ -453,7 +459,7 @@ public class GlobalAttributeEdit extends RequireLoginFirstAction
             }
         }
     }
- /*
+    /**
      * manages attribute to module/issue type mapping.
      */
